Force `ENV['environment']`
ActiveFedora is opinionated about how environment is set. Since we aren't an
engine, but we *are* setting `RAILS_ENV` via Hyrax, we run afoul of these
expectations and need to manually set the ENV to test to make `ActiveFedora`
happy.

Production apps will presumably setup the environment in a way that satisfies
these expectations.

See:
https://github.com/samvera/active_fedora/blob/2ff30dc1dc5b31f/lib/active_fedora.rb#L187-L209