Add support for ordered fields
The properties for ordered fields are added by an addition to `Schema`. Setters
are altered in `TuftsModel`, since they need to be changed after inclusion of
`Hyrax::BasicMetadata`.

Shared examples are added, and included in the various concrete
models. These tests are based on those in MIRA (`epigaea`), but refactored
slightly.