Include Hyrax Work/CollectionBehavior when available
In a Hyrax application, we want to automatically include the
`Hyrax::WorkBehaivor` and `Hyrax::CollectionBehavior` modules in their
appropriate models. This avoids applications needing to include them in
inherited models manually.

If these behaviors are not defined, then we continue to require and include
`Hyrax::CoreMetadata` and `Hyrax::BasicMetadata` manually.