Lock Hyrax support down to 2.0.x
We are concerned about concurrent applications attempting to share models across
minor versions of Hyrax, so we are locking support to `~> 2.0.0`.

Support for tiny versions within the minor series is left open, though we
recommend taking care when running two different versions of the Hyrax codebase
on applications that may load and save the same objects.