Add model setup macro
Adds a method intended for use in an initializer to define application layer
models. Using this with:

    Hyrax.config do |config|
      Tufts::Curation.setup_models!(configuration: config)
    end

Defines local model classes for each `Tufts::Curation` model, and registers with
the given config object (which should be a `Hyrax::Configuration`).

Indexers are defined automatically to use the default `Tufts::Curation::Indexer`
on each defined `TuftsModel` subclass.